refactor(world): change most values from ints to doubles
BREAKING CHANGE:
* Convert all positions and energy levels from `int` to `doubles`.
* Convert all usages of ID from `int` to `long`.

+++ src/model/model.cpp
@@ -1,6 +1,6 @@
 #include "model.h"
 
-Point::Point(int x, int y) : x{x}, y{y} {}
+Point::Point(double x, double y) : x{x}, y{y} {}
 
 bool operator == (const Point & a, const Point & b) {return a.x == b.x && a.y == b.y;}
 double operator / (const Point & a, const Point & b) {return sqrt(pow(b.x - a.x, 2.0) + pow(b.y - a.y, 2.0));}
