+++ src/model/object/entity/entity.cpp
@@ -1,6 +1,6 @@
 #include "entity.h"
 
-Entity::Entity(int x, int y, int energy) :
+Entity::Entity(double x, double y, double energy) :
     Object(x, y),
     energy{energy}
 {checkEnergy();}
@@ -9,9 +9,9 @@ void Entity::checkEnergy() {
     if (energy <= 0) throw std::range_error("Entity::checkEnergy(): Energy has to be positive");
 }
 
-int Entity::getEnergy() {return energy;}
+double Entity::getEnergy() {return energy;}
 
 void Entity::setPosition(Point p) { position = p;}
-void Entity::setPosition(int x, int y) { position = Point(x, y);}
+void Entity::setPosition(double x, double y) { position = Point(x, y);}
 
 Entity::Object::~Object() = default;
