+++ src/model/world/world.cpp
@@ -1,6 +1,6 @@
 #include "world.h"
 
-World::World(int WIDTH, int HEIGHT, int ENERGY) :
+World::World(double WIDTH, double HEIGHT, double ENERGY) :
     WIDTH{WIDTH},
     HEIGHT{HEIGHT},
     ENERGY{ENERGY}
@@ -30,7 +30,7 @@ void World::incEnergy(Entity * entity) {
     energy += entity->getEnergy();
 }
 
-int World::getEnergy() {
+double World::getEnergy() {
     return energy;
 }
 
@@ -72,7 +72,7 @@ std::list<Individual *> World::getCemetery() {
     return cemetery;
 }
 
-bool World::kill(int ID) {
+bool World::kill(long ID) {
     unsigned long index = 0;
     for (auto individual : individuals) {
         if (individual->getID() == ID) {
